[0293] - Migrate to separate class
